Removed numpy functions to encourage users to use jax instead.
